Make jobs parallelization configurable.  Add a -r[esume] flag to not clean.

+++ bin/llvm-bootstrap.sh
@@ -12,10 +12,24 @@ if [ ! -d llvm-project ]; then
 fi
 
 cd llvm-project || barf "Can't find llvm-project directory?!"
-rm -rf build
-mkdir build
+
+if [ "$1" = "-p" ]; then
+	shift;
+	_P=$1
+else
+	_P=8
+fi
+
+if [ "$1" = "-r" ]; then
+	echo "Resuming: Not cleaning first."
+	shift
+else
+	rm -rf build
+	mkdir build
+fi
 cd build
+
 cmake -DLLVM_ENABLE_PROJECTS='clang;clang-tools-extra;libcxx;libcxxabi;libunwind;lldb;compiler-rt;lld;polly' ../llvm || barf "Cmake step failed"
-make -j4 || barf "Unable to build llvm project sources"
+make -j${_P} || barf "Unable to build llvm project sources"
 sudo make install || barf "Installation failed"
 exit 0
